+++ lib/Models/ItemSearchProviders/IndexedItemSearchProvider.ts
@@ -2,7 +2,7 @@ import i18next from "i18next";
 import { action } from "mobx";
 import loadCsv from "../../Core/loadCsv";
 import loadJson5 from "../../Core/loadJson5";
-import { SearchParameterTraits } from "../../Traits/SearchableItemTraits";
+import { SearchParameterTraits } from "../../Traits/TraitsClasses/SearchableItemTraits";
 import ItemSearchProvider, {
   ItemSearchParameter,
   ItemSearchResult